Add parameter "--check_queue".
If provided, the application will check if the size of a certain queue
or queues exceeds the limit, and stop publishing until it is over this
limit. Format: --check_queue queuename=maxsize

Example:

    $ rmq_publish -d . -e "" -r "myqueue" --check_queue myqueue=10000

to stop publishing into `myqueue` until its length is over 10000
messages.

Can be used multiple times to add checks for multiple queues.